filter_persistence -> position_persistence
also lower it from 8 to 4

decrement only by 0.5
this way decrement is "as fast" as increment which increases either
odd or even by 1

+++ track.h
@@ -293,8 +293,8 @@ struct aircraft
   unsigned pos_rc; // Rc of last computed position
   double lat; // Coordinates obtained from CPR encoded data
   double lon; // Coordinates obtained from CPR encoded data
-  int pos_reliable_odd; // Number of good global CPRs, indicates position reliability
-  int pos_reliable_even;
+  float pos_reliable_odd; // Number of good global CPRs, indicates position reliability
+  float pos_reliable_even;
   int16_t traceWrittenForYesterday; // the permanent trace has been written for the previous day
   uint16_t mlatEPU;
   float gs_last_pos; // Save a groundspeed associated with the last position
